line chart: sub view for axis and grid
Sub-view is a region in a chart that allows content rendeing or adding
interactivity. In this instance, we are writing these sub-views in
Angular for drawing axis and grid.

This design choice absolves us from inventing our own layout system
within canvas while allowing us to add interactivity in line chart and
axis in the future.

Other implicit pros:
- do not have to render text in WebGL which is a bit challenging
- interactivity, which occurs in the main thread, does not have to be
  propagated to the worker based canvas renderer.